fix pour avoir qqch de plus python-friendly + fix du fichier test.py
